+++ server/src/main/java/org/allaymc/server/blockentity/component/BlockEntityEnderChestBaseComponentImpl.java
@@ -1,11 +1,11 @@
 package org.allaymc.server.blockentity.component;
 
 import lombok.Getter;
+import org.allaymc.api.block.action.SimpleBlockAction;
 import org.allaymc.api.blockentity.component.BlockEntityEnderChestBaseComponent;
 import org.allaymc.api.blockentity.initinfo.BlockEntityInitInfo;
 import org.allaymc.api.math.MathUtils;
 import org.allaymc.api.world.sound.SimpleSound;
-import org.cloudburstmc.protocol.bedrock.packet.BlockEventPacket;
 
 /**
  * @author IWareQ
@@ -23,21 +23,11 @@ public class BlockEntityEnderChestBaseComponentImpl extends BlockEntityBaseCompo
         this.viewersCount = count;
 
         if (viewersCount == 1) {
-            BlockEventPacket packet = new BlockEventPacket();
-            packet.setBlockPosition(position.toNetwork());
-            packet.setEventType(1);
-            packet.setEventData(2);
-            sendPacketToViewers(packet);
-
             position.dimension().addSound(MathUtils.center(position), SimpleSound.ENDER_CHEST_OPEN);
+            position.dimension().addBlockAction(position, SimpleBlockAction.OPEN);
         } else if (viewersCount == 0) {
-            BlockEventPacket packet = new BlockEventPacket();
-            packet.setBlockPosition(position.toNetwork());
-            packet.setEventType(1);
-            packet.setEventData(0);
-            sendPacketToViewers(packet);
-
             position.dimension().addSound(MathUtils.center(position), SimpleSound.ENDER_CHEST_CLOSE);
+            position.dimension().addBlockAction(position, SimpleBlockAction.CLOSE);
         }
     }
 }